refactor(constants): server
